refact :  clean up pass rules additioanl params
