+++ src/WorkflowsRules/Actions/ADF.php
@@ -3,11 +3,12 @@
 namespace Kanvas\Packages\WorkflowsRules\Actions;
 
 use Kanvas\Hengen\Hengen;
+use Kanvas\Packages\WorkflowsRules\Actions;
 use Kanvas\Packages\WorkflowsRules\Contracts\Interfaces\WorkflowsEntityInterfaces;
 use Phalcon\Di;
 use Throwable;
 
-class ADF extends Action
+class ADF extends Actions
 {
     /**
      * handle.
@@ -18,10 +19,11 @@ class ADF extends Action
      *
      * @return void
      */
-    public function handle(WorkflowsEntityInterfaces $entity, ...$args) : void
+    public function handle(WorkflowsEntityInterfaces $entity) : void
     {
         $response = null;
         $di = Di::getDefault();
+        $args = $entity->getRulesRelatedEntities();
 
         try {
             $transformer = Hengen::getTransformer(
@@ -46,10 +48,10 @@ class ADF extends Action
                 'data' => $transformer->getData()
             ]);
 
-            $this->setStatus(Action::SUCCESSFUL);
+            $this->setStatus(Actions::SUCCESSFUL);
         } catch (Throwable $e) {
             $this->setError($e->getMessage());
-            $this->setStatus(Action::FAIL);
+            $this->setStatus(Actions::FAIL);
         }
     }
 }
